+++ src/grepros/outputs.py
@@ -84,10 +84,14 @@ class SinkBase(object):
 class TextSinkMixin(object):
     """Provides message formatting as text."""
 
+    ## Default highlight wrappers if not color output
+    NOCOLOR_HIGHLIGHT_WRAPPERS = "**", "**"
+
 
     def __init__(self, args):
         """
         @param   args                       arguments object like argparse.Namespace
+        @param   args.COLOR                 "never" for not using colors in replacements
         @param   args.PRINT_FIELDS          message fields to use in output if not all
         @param   args.NOPRINT_FIELDS        message fields to skip in output
         @param   args.MAX_FIELD_LINES       maximum number of lines to output per field
@@ -97,12 +101,15 @@ class TextSinkMixin(object):
         @param   args.LINES_AROUND_MATCH    number of message lines around matched fields to output
         @param   args.MATCHED_FIELDS_ONLY   output only the fields where match was found
         @param   args.WRAP_WIDTH            character width to wrap message YAML output at
+        @param   args.MATCH_WRAPPER         string to wrap around matched values,
+                                            both sides if one value, start and end if more than one,
+                                            or no wrapping if zero values
         """
-        self._prefix   = ""    # Put before each message line (filename if grepping 1+ files)
-        self._wrapper  = None  # TextWrapper instance
-        self._patterns = {}    # {key: [(() if any field else ('nested', 'path'), re.Pattern), ]}
-        self._format_repls = {MatchMarkers.START: ConsolePrinter.HIGHLIGHT_START,
-                              MatchMarkers.END:   ConsolePrinter.HIGHLIGHT_END}
+        self._prefix       = ""    # Put before each message line (filename if grepping 1+ files)
+        self._wrapper      = None  # TextWrapper instance
+        self._patterns     = {}    # {key: [(() if any field else ('path', ), re.Pattern), ]}
+        self._format_repls = {}    # {text to replace if highlight: replacement text}
+        self._styles = collections.defaultdict(str)  # {label: ANSI code string}
 
         self._configure(args)
 
@@ -119,7 +126,7 @@ class TextSinkMixin(object):
                 start = self._args.START_LINE or 0
                 start = max(start, -len(lines)) - (start > 0)  # <0 to sanity, >0 to 0-base
                 lines = lines[start:start + (self._args.MAX_MESSAGE_LINES or len(lines))]
-                lines = lines and (lines[:-1] + [lines[-1] + ConsolePrinter.STYLE_RESET])
+                lines = lines and (lines[:-1] + [lines[-1] + self._styles["rst"]])
 
             if self._args.LINES_AROUND_MATCH and highlight:
                 spans, NUM = [], self._args.LINES_AROUND_MATCH
@@ -128,7 +135,7 @@ class TextSinkMixin(object):
                         spans.append([max(0, i - NUM), min(i + NUM + 1, len(lines))])
                     if MatchMarkers.END in l and spans:
                         spans[-1][1] = min(i + NUM + 1, len(lines))
-                lines = sum((lines[a:b - 1] + [lines[b - 1] + ConsolePrinter.STYLE_RESET]
+                lines = sum((lines[a:b - 1] + [lines[b - 1] + self._styles["rst"]]
                              for a, b in merge_spans(spans)), [])
 
             if self._prefix:
@@ -205,11 +212,12 @@ class TextSinkMixin(object):
 
 
     @classmethod
-    def make_nonconsole_args(cls, args):
+    def make_full_yaml_args(cls, args):
         """
-        Returns init arguments with all non-console arguments blanked to defaults.
+        Returns init arguments that provide message full YAMLs with no markers or colors.
 
         @param   args                       arguments object like argparse.Namespace
+        @param   args.COLOR                 set to "never" in result
         @param   args.PRINT_FIELDS          blanked to [] in result
         @param   args.NOPRINT_FIELDS        blanked to [] in result
         @param   args.MAX_FIELD_LINES       blanked to None in result
@@ -218,10 +226,13 @@ class TextSinkMixin(object):
         @param   args.MAX_MESSAGE_LINES     blanked to None in result
         @param   args.LINES_AROUND_MATCH    blanked to None in result
         @param   args.MATCHED_FIELDS_ONLY   blanked to False in result
+        @param   args.MATCH_WRAPPER         set to [""] in result
+        @param   args.WRAP_WIDTH            set to 120 in result
         """
         DEFAULTS = {"PRINT_FIELDS": [], "NOPRINT_FIELDS": [], "MAX_FIELD_LINES": None,
                     "START_LINE": None, "END_LINE": None, "MAX_MESSAGE_LINES": None,
-                    "LINES_AROUND_MATCH": None, "MATCHED_FIELDS_ONLY": False}
+                    "LINES_AROUND_MATCH": None, "MATCHED_FIELDS_ONLY": False,
+                    "COLOR": "never", "MATCH_WRAPPER": [""], "WRAP_WIDTH": 120}
         args = copy.deepcopy(args)
         for k, v in DEFAULTS.items(): setattr(args, k, v)
         return args
@@ -233,31 +244,38 @@ class TextSinkMixin(object):
         for key, vals in [("print", prints), ("noprint", noprints)]:
             self._patterns[key] = [(tuple(v.split(".")), wildcard_to_regex(v)) for v in vals]
 
-        HL0, HL1 = ConsolePrinter.HIGHLIGHT_START, ConsolePrinter.HIGHLIGHT_END
-        LL0, LL1 = ConsolePrinter.LOWLIGHT_START,  ConsolePrinter.LOWLIGHT_END
-        custom_widths = {
-            MatchMarkers.START:            len(HL0.replace(ConsolePrinter.STYLE_HIGHLIGHT, "")),
-            MatchMarkers.END:              len(HL1.replace(ConsolePrinter.STYLE_RESET,     "")),
-            ConsolePrinter.LOWLIGHT_START: 0,
-            ConsolePrinter.LOWLIGHT_END:   0,
-            ConsolePrinter.ERROR_START:    0,
-            ConsolePrinter.ERROR_END:      0,
-            ConsolePrinter.PREFIX_START:   0,
-            ConsolePrinter.PREFIX_END:     0,
-        }
-        width = ConsolePrinter.WIDTH if args.WRAP_WIDTH is None else args.WRAP_WIDTH
-        wrapargs = dict(width=width, max_lines=args.MAX_FIELD_LINES,
-                        placeholder="%s ...%s" % (LL0, LL1))
+        if "never" != args.COLOR:
+            self._styles.update({"hl0":  ConsolePrinter.STYLE_HIGHLIGHT,
+                                 "ll0":  ConsolePrinter.STYLE_LOWLIGHT,
+                                 "pfx0": ConsolePrinter.STYLE_SPECIAL, # Content line prefix start
+                                 "sep0": ConsolePrinter.STYLE_SPECIAL2})
+            self._styles.default_factory = lambda: ConsolePrinter.STYLE_RESET
+
+        WRAPS = args.MATCH_WRAPPER
+        if WRAPS is None and "never" == args.COLOR: WRAPS = self.NOCOLOR_HIGHLIGHT_WRAPPERS
+        WRAPS = ((WRAPS or [""]) * 2)[:2]
+        self._styles["hl0"] = self._styles["hl0"] + WRAPS[0]
+        self._styles["hl1"] = WRAPS[1] + self._styles["hl1"]
+
+        custom_widths = {MatchMarkers.START: len(WRAPS[0]), MatchMarkers.END:     len(WRAPS[1]),
+                         self._styles["ll0"]:            0, self._styles["ll1"]:  0,
+                         self._styles["pfx0"]:           0, self._styles["pfx1"]: 0,
+                         self._styles["sep0"]:           0, self._styles["sep1"]: 0}
+        wrapargs = dict(max_lines=args.MAX_FIELD_LINES,
+                        placeholder="%s ...%s" % (self._styles["ll0"], self._styles["ll1"]))
+        if args.WRAP_WIDTH is not None: wrapargs.update(width=args.WRAP_WIDTH)
         self._wrapper = TextWrapper(custom_widths=custom_widths, **wrapargs)
+        self._format_repls = {MatchMarkers.START: self._styles["hl0"],
+                              MatchMarkers.END:   self._styles["hl1"]}
 
 
 
 class ConsoleSink(SinkBase, TextSinkMixin):
     """Prints messages to console."""
 
-    META_LINE_TEMPLATE   = "{coloron}{sep} {line}{coloroff}"
-    MESSAGE_SEP_TEMPLATE = "{coloron}{sep}{coloroff}"
-    PREFIX_TEMPLATE      = "{coloron}{batch}{coloroff}{sepcoloron}{sep}{sepcoloroff}"
+    META_LINE_TEMPLATE   = "{ll0}{sep} {line}{ll1}"
+    MESSAGE_SEP_TEMPLATE = "{ll0}{sep}{ll1}"
+    PREFIX_TEMPLATE      = "{pfx0}{batch}{pfx1}{sep0}{sep}{sep1}"
     MATCH_PREFIX_SEP     = ":"    # Printed after bag filename for matched message lines
     CONTEXT_PREFIX_SEP   = "-"    # Printed after bag filename for context message lines
     SEP                  = "---"  # Prefix of message separators and metainfo lines
@@ -279,6 +297,10 @@ class ConsoleSink(SinkBase, TextSinkMixin):
         @param   args.MATCHED_FIELDS_ONLY   output only the fields where match was found
         @param   args.WRAP_WIDTH            character width to wrap message YAML output at
         """
+        if args.WRAP_WIDTH is None:
+            args = copy.deepcopy(args)
+            args.WRAP_WIDTH = ConsolePrinter.WIDTH
+
         super(ConsoleSink, self).__init__(args)
         TextSinkMixin.__init__(self, args)
 
@@ -290,8 +312,7 @@ class ConsoleSink(SinkBase, TextSinkMixin):
         batch = self._args.META and self.source.get_batch()
         if self._args.META and batch not in self._batch_meta:
             meta = self._batch_meta[batch] = self.source.format_meta()
-            kws = dict(coloron=ConsolePrinter.LOWLIGHT_START,
-                       coloroff=ConsolePrinter.LOWLIGHT_END, sep=self.SEP)
+            kws = dict(self._styles, sep=self.SEP)
             meta = "\n".join(x and self.META_LINE_TEMPLATE.format(**dict(kws, line=x))
                              for x in meta.splitlines())
             meta and ConsolePrinter.print(meta)
@@ -302,12 +323,9 @@ class ConsoleSink(SinkBase, TextSinkMixin):
         self._prefix = ""
         if self._args.LINE_PREFIX and self.source.get_batch():
             sep = self.MATCH_PREFIX_SEP if match else self.CONTEXT_PREFIX_SEP
-            kws = dict(coloron=ConsolePrinter.PREFIX_START, sep=sep,
-                       coloroff=ConsolePrinter.PREFIX_END, batch=self.source.get_batch(),
-                       sepcoloron=ConsolePrinter.SEP_START, sepcoloroff=ConsolePrinter.SEP_END)
+            kws = dict(self._styles, sep=sep, batch=self.source.get_batch())
             self._prefix = self.PREFIX_TEMPLATE.format(**kws)
-        kws = dict(coloron=ConsolePrinter.LOWLIGHT_START,
-                   coloroff=ConsolePrinter.LOWLIGHT_END, sep=self.SEP)
+        kws = dict(self._styles, sep=self.SEP)
         if self._args.META:
             meta = self.source.format_message_meta(topic, index, stamp, msg)
             meta = "\n".join(x and self.META_LINE_TEMPLATE.format(**dict(kws, line=x))
@@ -373,6 +391,9 @@ class HtmlSink(SinkBase, TextSinkMixin):
     TEMPLATE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
                                  "templates", "html.tpl")
 
+    ## Character wrap width for message YAML
+    WRAP_WIDTH = 120
+
     def __init__(self, args):
         """
         @param   args                    arguments object like argparse.Namespace
@@ -380,9 +401,12 @@ class HtmlSink(SinkBase, TextSinkMixin):
         @param   args.OUTFILE            name of HTML file to write,
                                          will add counter like .2 to filename if exists
         @param   args.OUTFILE_TEMPLATE   path to custom HTML template, if any
-        @param   args.WRAP_WIDTH         character width to wrap message YAML output at
         @param   args.VERBOSE            whether to print debug information
         """
+        args = copy.deepcopy(args)
+        args.WRAP_WIDTH = self.WRAP_WIDTH
+        args.COLOR = "always"
+
         super(HtmlSink, self).__init__(args)
         TextSinkMixin.__init__(self, args)
         self._queue    = queue.Queue()
@@ -392,8 +416,8 @@ class HtmlSink(SinkBase, TextSinkMixin):
         self._close_printed = False
         self._tag_repls = {MatchMarkers.START:            '<span class="match">',
                            MatchMarkers.END:              '</span>',
-                           ConsolePrinter.LOWLIGHT_START: '<span class="lowlight">',
-                           ConsolePrinter.LOWLIGHT_END:   '</span>'}
+                           ConsolePrinter.STYLE_LOWLIGHT: '<span class="lowlight">',
+                           ConsolePrinter.STYLE_RESET:    '</span>'}
         self._tag_rgx = re.compile("(%s)" % "|".join(map(re.escape, self._tag_repls)))
 
         self._format_repls.clear()
@@ -667,7 +691,8 @@ class SqliteSink(SinkBase, TextSinkMixin):
         @param   args.WRAP_WIDTH   character width to wrap message YAML output at
         @param   args.VERBOSE      whether to print debug information
         """
-        args = TextSinkMixin.make_nonconsole_args(args)
+        args = TextSinkMixin.make_full_yaml_args(args)
+
         super(SqliteSink, self).__init__(args)
         TextSinkMixin.__init__(self, args)
 
@@ -680,6 +705,7 @@ class SqliteSink(SinkBase, TextSinkMixin):
         # {"view": {topic: {typename: True}}, "table": {typename: {cols}}}
         self._schema = collections.defaultdict(dict)
 
+        self._format_repls.update({k: "" for k in self._format_repls})
         atexit.register(self.close)